feat: (en contra de mi voluntad) Se actualiza ApiExceptionHandler para que maneje correctamente los errores de autorización y autenticación, se actualiza el application.yaml para que todos los floats de los requests sean parseados a BigDecimal automáticamente, Se cambia a Integer los campos de reorderPoint y minStock de producto por mejor manejabilidad, se actualiza SecurityConfig para que maneje bien los errores de autorización y autenticación

+++ backend/WarehouseInventory/src/main/java/org/example/warehouseinventory/catalog/domain/entity/Product.java
@@ -38,8 +38,8 @@ public class Product extends AuditableEntity {
     @Embedded
     private Weight weight;
 
-    private BigDecimal minStockLevel; //helps to make an alert for restock
-    private BigDecimal reorderPoint; //help with ROP
+    private Integer minStockLevel; //helps to make an alert for restock
+    private Integer reorderPoint; //help with ROP
 
     @Enumerated(EnumType.STRING)
     private ProductCategory productCategory;
